# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VARUNA API",
    "version": "1.0.0",
    "description": "Vessel Attribution through Remote-sensing & Unified Navigational Analytics (SIH26143). Errors are RFC 9457 application/problem+json. Auth is via httpOnly cookies."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/detections/rejection-categories": {
      "get": {
        "tags": [
          "detections"
        ],
        "summary": "The rejection taxonomy a REJECT review must name. A category is usable as a labelled negative for the detector iff it names a sarClass; OPERATIONAL categories describe the workflow rather than the imagery and are deliberately excluded from training.",
        "responses": {
          "200": {
            "description": "Every category, with its kind, training class and description",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string",
                            "enum": [
                              "LOOK_ALIKE",
                              "OPERATIONAL"
                            ]
                          },
                          "sarClass": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "enum": [
                              "sea_surface",
                              "oil_spill",
                              "look_alike",
                              "ship",
                              "land"
                            ]
                          },
                          "description": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "label",
                          "kind",
                          "sarClass",
                          "description"
                        ]
                      }
                    },
                    "note": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items",
                    "note"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "RFC 9457 problem+json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      }
                    },
                    "attempted": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider": {
                            "type": "string"
                          },
                          "outcome": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "provider",
                          "outcome"
                        ]
                      }
                    },
                    "consequence": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/detections/{id}/review": {
      "post": {
        "tags": [
          "detections"
        ],
        "summary": "CONFIRM / REJECT / EDIT / REOPEN. The model output is immutable — an EDIT appends a version carrying the pre-edit geometry rather than overwriting it. REJECT requires BOTH a note (so the decision is reviewable) and a rejectionCategory (so it counts across cases and can become a labelled negative).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "CONFIRM",
                      "REJECT",
                      "EDIT",
                      "REOPEN"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 4000
                  },
                  "rejectionCategory": {
                    "type": "string",
                    "enum": [
                      "LOW_WIND",
                      "BIOGENIC_FILM",
                      "RAIN_CELL",
                      "SHIP_WAKE",
                      "INTERNAL_WAVE",
                      "SEA_ICE",
                      "LAND_OR_STRUCTURE",
                      "SENSOR_ARTEFACT",
                      "LOOK_ALIKE_UNCLASSIFIED",
                      "INSUFFICIENT_IMAGE_QUALITY",
                      "DUPLICATE",
                      "SUPERSEDED",
                      "OUT_OF_SCOPE"
                    ]
                  },
                  "geometry": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "enum": [
                          "Polygon"
                        ]
                      },
                      "coordinates": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "prefixItems": [
                              {
                                "type": "number",
                                "minimum": -180,
                                "maximum": 180
                              },
                              {
                                "type": "number",
                                "minimum": -90,
                                "maximum": 90
                              }
                            ]
                          },
                          "minItems": 4
                        },
                        "minItems": 1
                      }
                    },
                    "required": [
                      "type",
                      "coordinates"
                    ]
                  }
                },
                "required": [
                  "action"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reviewed. `trainingClass` is the SAR class this rejection contributes as a labelled negative, or null when it contributes none.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detectionId": {
                      "type": "string"
                    },
                    "reviewStatus": {
                      "type": "string"
                    },
                    "version": {
                      "type": "integer"
                    },
                    "areaKm2": {
                      "type": "number"
                    },
                    "geometryChanged": {
                      "type": "boolean"
                    },
                    "rejectionCategory": {
                      "type": "string",
                      "enum": [
                        "LOW_WIND",
                        "BIOGENIC_FILM",
                        "RAIN_CELL",
                        "SHIP_WAKE",
                        "INTERNAL_WAVE",
                        "SEA_ICE",
                        "LAND_OR_STRUCTURE",
                        "SENSOR_ARTEFACT",
                        "LOOK_ALIKE_UNCLASSIFIED",
                        "INSUFFICIENT_IMAGE_QUALITY",
                        "DUPLICATE",
                        "SUPERSEDED",
                        "OUT_OF_SCOPE"
                      ]
                    },
                    "trainingClass": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "sea_surface",
                        "oil_spill",
                        "look_alike",
                        "ship",
                        "land"
                      ]
                    }
                  },
                  "required": [
                    "detectionId",
                    "reviewStatus",
                    "version",
                    "areaKm2",
                    "geometryChanged"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "RFC 9457 problem+json",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      }
                    },
                    "attempted": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider": {
                            "type": "string"
                          },
                          "outcome": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "provider",
                          "outcome"
                        ]
                      }
                    },
                    "consequence": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "A REJECT without a note or without a category, or an EDIT without a geometry.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "path",
                          "message"
                        ]
                      }
                    },
                    "attempted": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider": {
                            "type": "string"
                          },
                          "outcome": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "provider",
                          "outcome"
                        ]
                      }
                    },
                    "consequence": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}